Fixed ChatBlock Mask for Player Updating thanks to Polar.

+++ src/com/oldscape/shared/model/AccessMaskType.java
@@ -61,7 +61,7 @@ public enum AccessMaskType {
     }
 
     public int getID() {
-        return type;
+        return (byte) type;
     }
 
 }
